finaprengine/requirements.txt
autopep8==1.4.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pymongo==3.9.0
six==1.12.0
SQLAlchemy==1.3.11
Werkzeug==0.16.0
